fastapi-master-api/pyproject.toml
[tool.poetry]
name = "fullstack-porfolio"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
SQLAlchemy = "^1.4.27"
uvicorn = "^0.15.0"
PyYAML = "^6.0"
python-dotenv = "^0.19.2"
pydantic = "^1.8.2"
psycopg2 = "^2.9.2"
PyJWT = "^2.3.0"
python-multipart = "^0.0.5"
passlib = "^1.7.4"
alembic = "^1.7.5"
importlib-resources = "^5.4.0"
bcrypt = "^3.2.0"
SQLAlchemy-Utils = "^0.38.1"
factory-boy = "^3.2.1"
pytz = "^2021.3"
pytest-html = "^3.1.1"
genbadge = "^1.0.6"
flake8-html = "^0.4.1"
pytest-cov = "^3.0.0"
pytest-freezegun = "^0.4.2"
black = "^21.12b0"
cryptography = "^36.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
faker-vehicle = "^0.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]
experimental-string-processing = 1

[tool.isort]
profile = "black"
filter_files = "1"
default_section = "THIRDPARTY"
known_django = ["django", "rest_framework"]
src_paths = "server"
sections = "FUTURE,STDLIB,THIRDPARTY,DJANGO,FIRSTPARTY,LOCALFOLDER"
skip = "migrations"
